w3c-reports/12092014/gtk/runned-tests/html/syntax/parsing/html5lib_tests24-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

html5lib Parser Test


FAIL html5lib_tests24.html 692c2dbacf18cb758f26a3d9e7d9add4356f9067 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html 614cc9827d3a0c5f91863dde5281dd0d97f64e6d null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html c517924583ee71b8e684c9ca1f2eed5e88139a39 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html 140a82fab878c139b388d159c511eb999fe2d8c7 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html a2ddddcccb652a6529daafd4153a0e12b6d5ca8c null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html da2e30a0b6577b608bf48bbd11a16ff832bc7e46 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html 66a5777f5453bd4b5161f00df02883b6d71f7cea null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_tests24.html c8d97f31b70f67005eeacc3c86ac29e577c3d0ed null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
